+++ client/operations/get_machine_parameters.go
@@ -56,7 +56,7 @@ type GetMachineParams struct {
 	  machine identifier
 
 	*/
-	ID int64
+	ID int32
 
 	timeout    time.Duration
 	Context    context.Context
@@ -86,13 +86,13 @@ func (o *GetMachineParams) SetContext(ctx context.Context) {
 }
 
 // WithID adds the id to the get machine params
-func (o *GetMachineParams) WithID(id int64) *GetMachineParams {
+func (o *GetMachineParams) WithID(id int32) *GetMachineParams {
 	o.SetID(id)
 	return o
 }
 
 // SetID adds the id to the get machine params
-func (o *GetMachineParams) SetID(id int64) {
+func (o *GetMachineParams) SetID(id int32) {
 	o.ID = id
 }
 
@@ -103,7 +103,7 @@ func (o *GetMachineParams) WriteToRequest(r runtime.ClientRequest, reg strfmt.Re
 	var res []error
 
 	// path param id
-	if err := r.SetPathParam("id", swag.FormatInt64(o.ID)); err != nil {
+	if err := r.SetPathParam("id", swag.FormatInt32(o.ID)); err != nil {
 		return err
 	}
