+++ client/operations/put_thermal_simulation_parameters.go
@@ -63,7 +63,7 @@ type PutThermalSimulationParams struct {
 	  simulation identifier
 
 	*/
-	ID int64
+	ID int32
 
 	timeout    time.Duration
 	Context    context.Context
@@ -104,13 +104,13 @@ func (o *PutThermalSimulationParams) SetThermalSimulation(thermalSimulation *mod
 }
 
 // WithID adds the id to the put thermal simulation params
-func (o *PutThermalSimulationParams) WithID(id int64) *PutThermalSimulationParams {
+func (o *PutThermalSimulationParams) WithID(id int32) *PutThermalSimulationParams {
 	o.SetID(id)
 	return o
 }
 
 // SetID adds the id to the put thermal simulation params
-func (o *PutThermalSimulationParams) SetID(id int64) {
+func (o *PutThermalSimulationParams) SetID(id int32) {
 	o.ID = id
 }
 
@@ -129,7 +129,7 @@ func (o *PutThermalSimulationParams) WriteToRequest(r runtime.ClientRequest, reg
 	}
 
 	// path param id
-	if err := r.SetPathParam("id", swag.FormatInt64(o.ID)); err != nil {
+	if err := r.SetPathParam("id", swag.FormatInt32(o.ID)); err != nil {
 		return err
 	}
